KAN-156 refactor: 어시스턴트 메뉴 위치 계산 커스텀 훅 useUpdatePosition 구현
