+++ src/app/components/editor/ai-assistant-interface/ManualModification.tsx
@@ -1,4 +1,4 @@
-import { ChangeEvent, RefObject, useEffect, useState } from 'react'
+import { ChangeEvent, RefObject, useState } from 'react'
 
 import { Editor } from '@tiptap/react'
 import { FeedbackFormData, FeedbackOptionType } from 'types/chatbot/chatbot'
@@ -6,6 +6,8 @@ import { ActionOptionType, EvaluateStateType, TextSelectionRangeType } from 'typ
 
 import FillButton from '@components/buttons/FillButton'
 import Portal from '@components/modal/Portal'
+
+import useUpdatePosition from '@hooks/editor/useUpdatePosition'
 
 import FeedbackOptionMenu from './menu/FeedbackOptionMenu'
 import PrimaryActionMenu from './menu/PrimaryActionMenu'
@@ -32,7 +34,8 @@ export default function ManualModification({
   onOptionClick,
   handleSubmitFeedback,
 }: ManualModificationProps) {
-  const [position, setPosition] = useState({ top: 0, left: 0 })
+  const position = useUpdatePosition(editor, selectionRef)
+
   const [feedbackInput, setFeedbackInput] = useState('')
   const [isShowFeedbackMenu, setIsShowFeedbackMenu] = useState(false)
   const [isShowFeedbackInput, setIsShowFeedbackInput] = useState(false)
@@ -68,44 +71,6 @@ export default function ManualModification({
   const handleBadFeedbackClick = () => {
     setIsShowFeedbackMenu(true)
   }
-
-  const updatePosition = () => {
-    if (!editor || !selectionRef?.current) return
-
-    const { to } = selectionRef.current
-    const view = editor.view
-
-    try {
-      const endCoords = view.coordsAtPos(to)
-      const editorRect = editor.view.dom.getBoundingClientRect()
-      setPosition({
-        top: endCoords.bottom,
-        left: Math.max(editorRect.left, endCoords.left),
-      })
-    } catch (error) {
-      console.error(error)
-    }
-  }
-
-  useEffect(() => {
-    updatePosition()
-  }, [editor, selectionRef])
-
-  useEffect(() => {
-    const handleScroll = () => {
-      updatePosition()
-    }
-
-    const editorDOM = editor?.view.dom
-    const editorContainer = editorDOM?.closest('.tiptap') || window
-    editorContainer.addEventListener('scroll', handleScroll)
-    window.addEventListener('scroll', handleScroll) // 전체 페이지 스크롤도 감지
-
-    return () => {
-      editorContainer.removeEventListener('scroll', handleScroll)
-      window.removeEventListener('scroll', handleScroll)
-    }
-  }, [editor, selectionRef])
 
   return (
     <Portal>
